Created base class file for all plotting tests
