+++ verticapy/tests_new/plotting/highcharts/machine_learning/test_highcharts_champion_challenger.py
@@ -14,45 +14,11 @@ OR CONDITIONS OF ANY KIND, either express or implied.
 See the  License for the specific  language governing
 permissions and limitations under the License.
 """
-# Pytest
-import pytest
-
-# Standard Python Modules
-
-
-# Other Modules
-
-
 # Verticapy
+from verticapy.tests_new.plotting.base_test_files import ChampionChallengerPlot
 
 
-# Testing variables
-
-
-class TestHighchartsMachineLearningChampionChallengerPlot:
+class TestHighchartsMachineLearningChampionChallengerPlot(ChampionChallengerPlot):
     """
     Testing different attributes of Champion CHallenger Plot
     """
-
-    @pytest.fixture(scope="class")
-    def plot_result(self, champion_challenger_plot):
-        """
-        Create a champion challenger plot using AutoML
-        """
-        return champion_challenger_plot.plot()
-
-    @pytest.fixture(autouse=True)
-    def result(self, plot_result):
-        """
-        Get the plot results
-        """
-        self.result = plot_result
-
-    def test_properties_output_type(self, plotting_library_object):
-        """
-        Test if correct object created
-        """
-        # Arrange
-        # Act
-        # Assert - checking if correct object created
-        assert isinstance(self.result, plotting_library_object), "Wrong object created"
