+++ verticapy/tests_new/plotting/highcharts/machine_learning/test_highcharts_regression_tree.py
@@ -14,53 +14,10 @@ OR CONDITIONS OF ANY KIND, either express or implied.
 See the  License for the specific  language governing
 permissions and limitations under the License.
 """
-# Pytest
-import pytest
-
-# Standard Python Modules
-
-# Verticapy
-from verticapy.learn.tree import DecisionTreeRegressor
-from verticapy.tests_new.plotting.conftest import BasicPlotTests
-
-# Testing variables
-COL_NAME_1 = "0"
-COL_NAME_2 = "1"
+from verticapy.tests_new.plotting.base_test_files import LearningRegressionTreePlot
 
 
-class TestHighchartsMachineLearningRegressionTreePlot(BasicPlotTests):
+class TestHighchartsMachineLearningRegressionTreePlot(LearningRegressionTreePlot):
     """
     Testing different attributes of Regression Tree plot
     """
-
-    @pytest.fixture(autouse=True)
-    def model(self, schema_loader, dummy_dist_vd):
-        """
-        Load test model
-        """
-        model = DecisionTreeRegressor(name=f"{schema_loader}.model_titanic")
-        x_col = COL_NAME_1
-        y_col = COL_NAME_2
-        model.fit(dummy_dist_vd, x_col, y_col)
-        self.model = model
-        yield
-        model.drop()
-
-    @property
-    def cols(self):
-        """
-        Store labels for X,Y,Z axis to check.
-        """
-        return [
-            COL_NAME_1,
-            COL_NAME_2,
-        ]
-
-    def create_plot(self):
-        """
-        Create the plot
-        """
-        return (
-            self.model.plot,
-            {},
-        )
