+++ verticapy/tests_new/plotting/highcharts/test_highcharts_pie.py
@@ -14,50 +14,14 @@ OR CONDITIONS OF ANY KIND, either express or implied.
 See the  License for the specific  language governing
 permissions and limitations under the License.
 """
-# Pytest
-import pytest
-
-# Standard Python Modules
-
-# Other Modules
-
-
 # Vertica
+from verticapy.tests_new.plotting.base_test_files import VDCPiePlot, NestedVDFPiePlot
 
 
-# Testing variables
-COL_NAME = "check 1"
-COL_NAME_2 = "check 2"
-all_elements = {"1", "0", "A", "B", "C"}
-
-
-class TestHighchartsVDCPiePlot:
+class TestHighchartsVDCPiePlot(VDCPiePlot):
     """
     Testing different attributes of Pie plot on a vDataColumn
     """
-
-    @pytest.fixture(scope="class")
-    def plot_result(self, dummy_vd):
-        """
-        Create a pie plot for vDataColumn
-        """
-        return dummy_vd[COL_NAME].pie()
-
-    @pytest.fixture(autouse=True)
-    def result(self, plot_result):
-        """
-        Get the plot results
-        """
-        self.result = plot_result
-
-    def test_properties_output_type_for(self, plotting_library_object):
-        """
-        Test if correct object created
-        """
-        # Arrange
-        # Act
-        # Assert - checking if correct object created
-        assert isinstance(self.result, plotting_library_object), "Wrong object created"
 
     def test_plot_type_wedges(
         self,
@@ -70,51 +34,11 @@ class TestHighchartsVDCPiePlot:
         # Assert - check value corresponding to 0s
         assert len(self.result.data_temp[0].data) > 1
 
-    @pytest.mark.parametrize("kind, max_cardinality", [("donut", 2), ("rose", 4)])
-    def test_properties_output_type_for_all_options(
-        self,
-        dummy_vd,
-        plotting_library_object,
-        kind,
-        max_cardinality,
-    ):
-        """
-        Test different kind and max-cardinality options
-        """
-        # Arrange
-        # Act
-        result = dummy_vd[COL_NAME].pie(kind=kind, max_cardinality=max_cardinality)
-        # Assert - checking if correct object created
-        assert isinstance(result, plotting_library_object), "Wrong object created"
 
-
-class TestHighchartsNestedVDFPiePlot:
+class TestHighchartsNestedVDFPiePlot(NestedVDFPiePlot):
     """
     Testing different attributes of Pie plot on a vDataFrame
     """
-
-    @pytest.fixture(scope="class")
-    def plot_result_2(self, dummy_vd):
-        """
-        Create a pie plot for vDataFrame
-        """
-        return dummy_vd.pie([COL_NAME, COL_NAME_2])
-
-    @pytest.fixture(autouse=True)
-    def result(self, plot_result_2):
-        """
-        Get the plot results
-        """
-        self.result = plot_result_2
-
-    def test_properties_type(self, plotting_library_object):
-        """
-        Test if correct object created
-        """
-        # Arrange
-        # Act
-        # Assert - checking if correct object created
-        assert isinstance(self.result, plotting_library_object), "Wrong object created"
 
     def test_plot_type_wedges(
         self,
